+++ client/go/.gitignore
@@ -1,0 +1,2 @@
+pkg/
+*.sum
