+++ client/go/src/github.com/vespa-engine/vespa/cmd/root.go
@@ -1,0 +1,30 @@
+package cmd
+
+import (
+    "github.com/spf13/cobra"
+)
+
+var (
+	// Used for flags.
+	cfgFile     string
+	userLicense string
+
+	rootCmd = &cobra.Command{
+		Use:   "vespa",
+		Short: "A command-line tool for working with Vespa instances",
+		Long: `TO
+DO`,
+	}
+)
+
+// Execute executes the root command.
+func Execute() error {
+	return rootCmd.Execute()
+}
+
+func init() {
+	cobra.OnInitialize(initConfig)
+}
+
+func initConfig() {
+}
