+++ client/go/src/github.com/vespa-engine/vespa/go.mod
@@ -1,0 +1,5 @@
+module github.com/vespa-engine/vespa
+
+go 1.16
+
+require github.com/spf13/cobra v1.2.1 // indirect
