+++ client/go/src/github.com/vespa-engine/vespa/main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+  "github.com/vespa-engine/vespa/cmd"
+)
+
+func main() {
+  cmd.Execute()
+}
