--- client/gobuild.sh
@@ -1,1 +1,0 @@
-go build -o target/vespa src/main/go/vespa.go
